Rename Choreo Jaeger Reporter to Choreo Span Exporter
